# docs/apispec.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "MyoHanMeeting Backend",
        "description": "유기묘 입양 서비스 묘한만남의 백엔드 서버 문서입니다.",
        "contact": {
            "email": "bomlee427@gmail.com"
        },
        "version": "v1"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Generated server url"
        }
    ],
...
    "paths": {
...
        "/v1/adoption/notices": {
            "get": {
                "tags": [
                    "Adopt Notice"
                ],
                "summary": "분양공고 목록조회",
                "description": "검색 조건에 따른 분양 공고 목록을 조회합니다.",
                "operationId": "getNoticeList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "페이지를 설정합니다.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 0
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "리미트를 설정합니다.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int32",
                            "default": 100
                        }
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "제목으로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "content",
                        "in": "query",
                        "description": "내용으로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "authorName",
                        "in": "query",
                        "description": "작성자명(닉네임)으로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "noticeStatus",
                        "in": "query",
                        "description": "공고의 상태로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "catName",
                        "in": "query",
                        "description": "고양이 이름으로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "catSpecies",
                        "in": "query",
                        "description": "고양이 종으로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "catAge",
                        "in": "query",
                        "description": "고양이 나이로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "catSex",
                        "in": "query",
                        "description": "고양이 성별로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "neutered",
                        "in": "query",
                        "description": "중성화 여부로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "보호소 위치로 검색합니다. 지역 코드는 아래와 같습니다.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "description": "City Enum Value",
                            "enum": [
                                "SEOUL",
                                "SEJONG",
                                "BUSAN",
                                "DAEGU",
                                "INCHEON",
                                "GWANGJU",
                                "ULSAN",
                                "DAEJEON",
                                "GYEONGGI",
                                "GANGWON",
                                "CHUNGCHEONG_BUK",
                                "CHUNGCHEONG_NAM",
                                "JEOLLA_BUK",
                                "JEOLLA_NAM",
                                "GYEONGSANG_BUK",
                                "GYEONGSANG_NAM",
                                "JEJU"
                            ]
                        }
                    },
                    {
                        "name": "shelterName",
                        "in": "query",
                        "description": "보호소 이름으로 검색합니다.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "정렬 기준을 선택합니다.<br><br>, 를 구분자로 항목 앞에 -를 붙이면 내림차순, 붙이지 않으면 오름차순 정렬입니다.<br><br>정렬 가능한 항목은 다음과 같습니다: createdAt(작성일), applicationCount(신청수), commentCount(댓글수)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "-createdAt,applicationCount"
                    }
                ],
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDtoListAdoptNoticeSummary"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Adopt Notice"
                ],
                "summary": "분양공고 작성",
                "description": "분양공고를 작성합니다.",
                "operationId": "createNotice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adoptNoticeForm"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "작성 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDto"
                                },
                                "example": {
                                    "status": "200 OK",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SUCCESS",
                                    "data": {
                                        "noticeId": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증이 필요한 URI입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "401 UNAUTHORIZED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "AUTHORIZATION REQUIRED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/v1/adoption/notices/comments": {
            "post": {
                "tags": [
                    "Adopt Notice Comment"
                ],
                "summary": "분양공고 댓글 작성",
                "description": "분양공고 댓글을 작성합니다.",
                "operationId": "createNoticeComment",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/adoptNoticeCommentForm"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "작성 성공",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDto"
                                },
                                "example": {
                                    "status": "200 OK",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SUCCESS",
                                    "data": {
                                        "noticeCommentId": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "인증이 필요한 URI입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "401 UNAUTHORIZED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "AUTHORIZATION REQUIRED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "요청한 자원을 찾을 수 없습니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "404 NOT_FOUND",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "NO RESOURCE FOUND",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
        "/v1/adoption/notices/{noticeId}/comments": {
            "get": {
                "tags": [
                    "Adopt Notice Comment"
                ],
                "summary": "분양공고 댓글 조회",
                "description": "특정 분양공고에 달린 댓글을 조회합니다.",
                "operationId": "getNoticeCommentList",
                "parameters": [
                    {
                        "name": "noticeId",
                        "in": "path",
                        "description": "댓글을 조회하고자 하는 공고의 id입니다.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "잘못된 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "400 BAD_REQUEST",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "INVALID PARAMETER",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "서버측 오류입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "500 INTERNAL_SERVER_ERROR",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "SERVER ERROR",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CommonResponseDtoListAdoptNoticeComment"
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "잘못된 메서드 또는 URI 요청입니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "405 METHOD_NOT_ALLOWED",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "METHOD_NOT_ALLOWED",
                                    "debugMessage": ""
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "요청한 자원을 찾을 수 없습니다.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                },
                                "example": {
                                    "status": "404 NOT_FOUND",
                                    "timestamp": "2023-06-10T09:19:08.550Z",
                                    "message": "NO RESOURCE FOUND",
                                    "debugMessage": ""
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "description": "HTTP 상태 코드입니다.",
                        "enum": [
                            "200 OK",
                            "400 BAD_REQUEST",
                            "401 UNAUTHORIZED",
                            "403 FORBIDDEN",
                            "404 NOT_FOUND",
                            "405 METHOD_NOT_ALLOWED",
                            "500 INTERNAL_SERVER_ERROR"
                        ]
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "description": "응답에 대한 요약을 나타냅니다."
                    },
                    "debugMessage": {
                        "type": "string",
                        "description": "오류에 관한 상세 디버깅 메시지를 나타냅니다."
                    }
                }
            },
            "CommonResponseDto": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "200 OK"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "example": "SUCCESS"
                    },
                    "data": {
                        "type": "object"
                    }
                }
            },
...
            "adoptNoticeForm": {
                "type": "object",
                "properties": {
                    "cat": {
                        "$ref": "#/components/schemas/Cat"
                    },
                    "thumbnailId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "catPictures": {
                        "type": "array",
                        "items": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    "title": {
                        "type": "string"
                    },
                    "content": {
                        "type": "string"
                    },
                    "shelter": {
                        "$ref": "#/components/schemas/Shelter"
                    }
                }
            },
            "adoptNoticeCommentForm": {
                "type": "object",
                "properties": {
                    "noticeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "content": {
                        "type": "string"
                    }
                }
            },
...
            "AdoptNoticeSummary": {
                "type": "object",
                "properties": {
                    "noticeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "noticeTitle": {
                        "type": "string"
                    },
                    "noticeStatus": {
                        "type": "string"
                    },
                    "thumbnail": {
                        "type": "string"
                    },
                    "authorName": {
                        "type": "string"
                    },
                    "catName": {
                        "type": "string"
                    },
                    "catSpecies": {
                        "type": "string"
                    },
                    "shelterCity": {
                        "type": "string"
                    },
                    "applicationCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "commentCount": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
            "CommonResponseDtoListAdoptNoticeSummary": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "200 OK"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "example": "SUCCESS"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AdoptNoticeSummary"
                        }
                    }
                }
            },
            "CommonResponseDtoListAdoptNoticeComment": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "200 OK"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "message": {
                        "type": "string",
                        "example": "SUCCESS"
                    },
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AdoptNoticeComment"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Authorization": {
                "type": "http",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}